Refactor zsl.unittest to zsl.testing as requested in a code review
