+++ tests/resource/guarded_resource_test.py
@@ -23,8 +23,8 @@ from zsl.resource.guard import (guard, GuardedMixin, Access,
                                 PolicyViolationError, ResourcePolicy)
 
 from tests.resource.resource_test_helper import test_settings
-from zsl.unittest.test_utils import parent_module
-from zsl.unittest.http import json_loads, HTTPTestCase
+from zsl.testing.test_utils import parent_module
+from zsl.testing.http import json_loads, HTTPTestCase
 
 TEST_VALUE_CREATED = 'created'
 TEST_VALUE_READ = 'read'
